requirements.txt.txt
discord.py
icalendar
pytz

